API call for Question and Answer moved to back-end server
