+++ server/controllers/index.js
@@ -1,5 +1,6 @@
 exports.review = require('./review.js');
-exports.rating = require('./rating.js')
-exports.image = require('./image.js')
-exports.itemDetails = require('./itemDetails.js')
-exports.relatedID = require('./relatedID.js')
+exports.rating = require('./rating.js');
+exports.image = require('./image.js');
+exports.itemDetails = require('./itemDetails.js');
+exports.relatedID = require('./relatedID.js');
+exports.questionAnswer = require('./questionAnswer.js');
