+++ server/routes.js
@@ -1,8 +1,6 @@
 require("dotenv").config();
 const router = require("express").Router();
 const controller = require('./controllers/')
-// router.route('/getItemDetails')
-//   .get(controller.itemDetails.getItemDetails)
 
 router.post('/reviews', controller.review.get_product_reviews)
 router.post('/metadata', controller.review.get_metadata)
@@ -13,7 +11,6 @@ router.get('/relateItemsID', controller.relatedID.getRelatedItemsID)
 router.get('/getRating', controller.rating.getRating)
 router.get('/getImageAndPrice', controller.image.getImageAndPrice)
 router.get('/getItemDetails', controller.itemDetails.getItemDetails)
-
-
+router.get('/getProductQA', controller.questionAnswer.getProductQA)
 
 module.exports = router;
